Move ModelConfiguration to top level
