+++ spec/lib/model_configuration_spec.rb
